fix(security): break circular deps, fix test mocks, add validatedFetch
- Move FetchResult/ExtractedReferral types from fetcher.ts to types.ts
  to break 4 circular dependencies in research-agent module
- Update page-fetcher, reddit-fetcher, referral-extractor, api-fetchers
  to import types from types.ts instead of fetcher.ts
- Add validatedFetch wrapper in security.ts for Codacy SSRF compliance
- Mock security module in reward-scraper tests to fix SSRF validation
- Import RESEARCH_SOURCES from barrel in research-agent test
