+++ .agents/metrics.jsonl
@@ -1,3 +1,4 @@
 {"timestamp": "2026-06-06T06:14:38Z", "agent": "opencode", "task": "Commit, push, and fix all CI checks for PR #425 (fix CLOUDFLARE_WORKER_HOST rename). Added Rationalizations/Red Flags to 9 SKILL.md files, excluded turnstile-spin templates from vitest/prettier, added .codacy.yml to suppress false-positive security findings on Cloudflare-documented test secrets, and updated root-allowlists in pre-commit-hook.sh and check-directory-organization.sh for .codacy.yml/.yaml. Result: 22/22 GH Actions pass (Validate Skills, Format Check, Quality Gate, Run Tests, Unit Tests, Codacy, etc.). The only remaining failure is Workers Builds (pre-existing, requires manual repo-admin actions documented in PR body: set WORKER_HOST variable and run Vectorize Index Setup workflow).", "status": "completed", "duration_seconds": 900}
 {"timestamp":"2026-06-06T09:44:46.580Z","agent":"jules","task":"Add secrets.required to wrangler.jsonc (Cloudflare 2026 best practice)","status":"completed","duration_seconds":180}
 {"timestamp": "2026-06-05T14:30:00Z", "agent": "opencode", "task": "Fix WORKER_HOST multi-env resolution (#423 round 2) + document Cloudflare Builds architecture", "status": "completed", "duration_seconds": 0}
+{"timestamp": "2026-06-09T09:57:40Z", "agent": "opencode", "task": "CI/CD swarm audit: fixed 3 critical bugs (deploy-production worker-host.sh, rollback.yml hardcoded URL, canary.yml missing checkout), upgraded setup-node v4→v6 across 13 workflows, closed 6 stale issues, created PR #442", "status": "completed", "duration_seconds": 600}
