+++ .github/ci-status/ci-status.json
@@ -1,6 +1,6 @@
 {
   "status": "passing",
-  "last_run": "2026-06-09T00:00:00Z",
+  "last_run": "2026-06-09T10:31:14Z",
   "failing_jobs": [],
-  "workflow_url": ""
+  "workflow_url": "https://github.com/do-ops885/do-deal-relay/actions/runs/27200162210"
 }
